Take screenshot from xcrun commands instead of idb
