+++ maestro-ios/src/main/java/ios/xcrun/Simctl.kt
@@ -6,6 +6,7 @@ import maestro.utils.MaestroTimer
 import net.harawata.appdirs.AppDirsFactory
 import okio.buffer
 import okio.source
+import util.CommandLineUtils
 import java.io.File
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
